+++ src/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "textile-js" library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Added
+- [LIBRARY] Add a `tokenize( textileString, options )` member, to get the JSONML tree without a global 'html' node.
+
 
 ## [2.0.101] - 20200204
 ### Added